Pivot Avx512Bf16 surface to Vector<ushort> per maintainer guidance
@tannergooding's review on the API proposal (dotnet/runtime#129323):
the strongly-typed BFloat16 element shape requires extensive JIT/VM/
Debugger/ABI work to make Vector128<BFloat16>.IsSupported report true,
which is the same reason F16C/FP16 have not landed yet. The cheaper
path he endorsed is to ship the initial signatures as Vector*<ushort>
(raw bf16 bit pattern) and add Vector*<BFloat16> overloads later when
the primitive-type plumbing is ready.

This commit applies that pivot:
- Avx512Bf16.cs / Avx512Bf16.PlatformNotSupported.cs: all BFloat16
  parameter / return types switched to ushort.
- System.Runtime.Intrinsics.cs (ref): same.

Method names (ConvertToBFloat16) are unchanged — they refer to the
operation, not the parameter type. Existing JIT plumbing
(InstructionSetDesc, codeman gate, lookupInstructionSet dispatch,
HARDWARE_INTRINSIC rows, CPUID detection, configs) is unaffected.

+++ src/libraries/System.Private.CoreLib/src/System/Runtime/Intrinsics/X86/Avx512Bf16.cs
@@ -1,12 +1,16 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Numerics;
 using System.Runtime.CompilerServices;
 
 namespace System.Runtime.Intrinsics.X86
 {
     /// <summary>Provides access to X86 AVX512-BF16 hardware instructions via intrinsics.</summary>
+    /// <remarks>
+    /// BF16 operands are exposed as <see cref="Vector128{T}"/> / <see cref="Vector256{T}"/> / <see cref="Vector512{T}"/>
+    /// of <see cref="ushort"/> (raw bf16 bit pattern). Strongly-typed <c>BFloat16</c> overloads are intended as a
+    /// follow-up once the <c>System.Numerics.BFloat16</c> primitive is wired through the JIT/VM SIMD pipeline.
+    /// </remarks>
     [Intrinsic]
     [CLSCompliant(false)]
     public abstract class Avx512Bf16 : Avx512F
@@ -25,8 +29,6 @@ namespace System.Runtime.Intrinsics.X86
             internal X64() { }
 
             /// <summary>Gets a value that indicates whether the APIs in this class are supported.</summary>
-            /// <value><see langword="true" /> if the APIs are supported; otherwise, <see langword="false" />.</value>
-            /// <remarks>A value of <see langword="false" /> indicates that the APIs will throw <see cref="PlatformNotSupportedException" />.</remarks>
             public static new bool IsSupported { get => IsSupported; }
         }
 
@@ -34,21 +36,21 @@ namespace System.Runtime.Intrinsics.X86
         ///   <para>__m512 _mm512_dpbf16_ps (__m512 src, __m512bh a, __m512bh b)</para>
         ///   <para>  VDPBF16PS zmm1, zmm2, zmm3/m512</para>
         /// </summary>
-        public static Vector512<float> MultiplyWideningAndAdd(Vector512<float> addend, Vector512<BFloat16> left, Vector512<BFloat16> right) => MultiplyWideningAndAdd(addend, left, right);
+        public static Vector512<float> MultiplyWideningAndAdd(Vector512<float> addend, Vector512<ushort> left, Vector512<ushort> right) => MultiplyWideningAndAdd(addend, left, right);
 
         /// <summary>
         ///   <para>__m512bh _mm512_cvtne2ps_pbh (__m512 a, __m512 b)</para>
         ///   <para>  VCVTNE2PS2BF16 zmm1, zmm2, zmm3/m512</para>
-        ///   <para>Round-to-nearest-even conversion of two packed FP32 vectors into a single packed BF16 vector.</para>
+        ///   <para>Round-to-nearest-even conversion of two packed FP32 vectors into a single packed BF16 vector (returned as <see cref="Vector512{T}"/> of <see cref="ushort"/>).</para>
         /// </summary>
-        public static Vector512<BFloat16> ConvertToBFloat16(Vector512<float> lower, Vector512<float> upper) => ConvertToBFloat16(lower, upper);
+        public static Vector512<ushort> ConvertToBFloat16(Vector512<float> lower, Vector512<float> upper) => ConvertToBFloat16(lower, upper);
 
         /// <summary>
         ///   <para>__m256bh _mm512_cvtneps_pbh (__m512 a)</para>
         ///   <para>  VCVTNEPS2BF16 ymm1, zmm2/m512</para>
-        ///   <para>Round-to-nearest-even conversion of a packed FP32 vector into a packed BF16 vector half-width.</para>
+        ///   <para>Round-to-nearest-even conversion of a packed FP32 vector into a half-width packed BF16 vector.</para>
         /// </summary>
-        public static Vector256<BFloat16> ConvertToBFloat16(Vector512<float> value) => ConvertToBFloat16(value);
+        public static Vector256<ushort> ConvertToBFloat16(Vector512<float> value) => ConvertToBFloat16(value);
 
         /// <summary>Provides access to the x86 AVX512-BF16+VL hardware instructions via intrinsics.</summary>
         [Intrinsic]
@@ -57,45 +59,25 @@ namespace System.Runtime.Intrinsics.X86
             internal VL() { }
 
             /// <summary>Gets a value that indicates whether the APIs in this class are supported.</summary>
-            /// <value><see langword="true" /> if the APIs are supported; otherwise, <see langword="false" />.</value>
-            /// <remarks>A value of <see langword="false" /> indicates that the APIs will throw <see cref="PlatformNotSupportedException" />.</remarks>
             public static new bool IsSupported { get => IsSupported; }
 
-            /// <summary>
-            ///   <para>__m128 _mm_dpbf16_ps (__m128 src, __m128bh a, __m128bh b)</para>
-            ///   <para>  VDPBF16PS xmm1, xmm2, xmm3/m128</para>
-            /// </summary>
-            public static Vector128<float> MultiplyWideningAndAdd(Vector128<float> addend, Vector128<BFloat16> left, Vector128<BFloat16> right) => MultiplyWideningAndAdd(addend, left, right);
+            /// <summary>__m128 _mm_dpbf16_ps (__m128 src, __m128bh a, __m128bh b) — VDPBF16PS xmm</summary>
+            public static Vector128<float> MultiplyWideningAndAdd(Vector128<float> addend, Vector128<ushort> left, Vector128<ushort> right) => MultiplyWideningAndAdd(addend, left, right);
 
-            /// <summary>
-            ///   <para>__m256 _mm256_dpbf16_ps (__m256 src, __m256bh a, __m256bh b)</para>
-            ///   <para>  VDPBF16PS ymm1, ymm2, ymm3/m256</para>
-            /// </summary>
-            public static Vector256<float> MultiplyWideningAndAdd(Vector256<float> addend, Vector256<BFloat16> left, Vector256<BFloat16> right) => MultiplyWideningAndAdd(addend, left, right);
+            /// <summary>__m256 _mm256_dpbf16_ps (__m256 src, __m256bh a, __m256bh b) — VDPBF16PS ymm</summary>
+            public static Vector256<float> MultiplyWideningAndAdd(Vector256<float> addend, Vector256<ushort> left, Vector256<ushort> right) => MultiplyWideningAndAdd(addend, left, right);
 
-            /// <summary>
-            ///   <para>__m128bh _mm_cvtne2ps_pbh (__m128 a, __m128 b)</para>
-            ///   <para>  VCVTNE2PS2BF16 xmm1, xmm2, xmm3/m128</para>
-            /// </summary>
-            public static Vector128<BFloat16> ConvertToBFloat16(Vector128<float> lower, Vector128<float> upper) => ConvertToBFloat16(lower, upper);
+            /// <summary>__m128bh _mm_cvtne2ps_pbh (__m128 a, __m128 b) — VCVTNE2PS2BF16 xmm</summary>
+            public static Vector128<ushort> ConvertToBFloat16(Vector128<float> lower, Vector128<float> upper) => ConvertToBFloat16(lower, upper);
 
-            /// <summary>
-            ///   <para>__m256bh _mm256_cvtne2ps_pbh (__m256 a, __m256 b)</para>
-            ///   <para>  VCVTNE2PS2BF16 ymm1, ymm2, ymm3/m256</para>
-            /// </summary>
-            public static Vector256<BFloat16> ConvertToBFloat16(Vector256<float> lower, Vector256<float> upper) => ConvertToBFloat16(lower, upper);
+            /// <summary>__m256bh _mm256_cvtne2ps_pbh (__m256 a, __m256 b) — VCVTNE2PS2BF16 ymm</summary>
+            public static Vector256<ushort> ConvertToBFloat16(Vector256<float> lower, Vector256<float> upper) => ConvertToBFloat16(lower, upper);
 
-            /// <summary>
-            ///   <para>__m128bh _mm_cvtneps_pbh (__m128 a)</para>
-            ///   <para>  VCVTNEPS2BF16 xmm1, xmm2/m128</para>
-            /// </summary>
-            public static Vector128<BFloat16> ConvertToBFloat16(Vector128<float> value) => ConvertToBFloat16(value);
+            /// <summary>__m128bh _mm_cvtneps_pbh (__m128 a) — VCVTNEPS2BF16 xmm</summary>
+            public static Vector128<ushort> ConvertToBFloat16(Vector128<float> value) => ConvertToBFloat16(value);
 
-            /// <summary>
-            ///   <para>__m128bh _mm256_cvtneps_pbh (__m256 a)</para>
-            ///   <para>  VCVTNEPS2BF16 xmm1, ymm2/m256</para>
-            /// </summary>
-            public static Vector128<BFloat16> ConvertToBFloat16(Vector256<float> value) => ConvertToBFloat16(value);
+            /// <summary>__m128bh _mm256_cvtneps_pbh (__m256 a) — VCVTNEPS2BF16 xmm from ymm</summary>
+            public static Vector128<ushort> ConvertToBFloat16(Vector256<float> value) => ConvertToBFloat16(value);
         }
     }
 }
